[REFACTOR]: split RDSButton into View/Types/Protocols/Layouts for clarity and scalability
