+++ Sources/RDSUI/Components/Buttons/RDSButton.swift
@@ -7,111 +7,6 @@
 
 import SwiftUI
 
-// MARK: - RDSButton Public API
-
-/// Visual variant of the button.
-/// - Primary: Solid background + border, selection color for text/icon.
-/// - Secondary: Outline-like; text/icon/border share color; background is white/clear.
-/// - Alternative: Solid background + border, selection color for text/icon.
-public enum RDSButtonVariant: Equatable {
-    case primary
-    case secondary
-    case alternative
-}
-
-/// Visual type inside a variant.
-/// - default: Standard filled/outlined button depending on variant.
-/// - textLink: Text-only style as defined by design system rules.
-public enum RDSButtonType: Equatable {
-    case `default`
-    case textLink
-}
-
-/// Fixed layout configurations (width/height and icon placement).
-/// Only `.iconLeft` and `.iconRight` render an icon if provided.
-public enum RDSButtonSize: Equatable {
-    case `default`    // 135x40, no icon
-    case large        // 151x56, no icon
-    case small        // 108x30, no icon
-    case iconLeft     // 148x40, icon leading
-    case iconRight    // 148x40, icon trailing
-}
-
-/// External button state as specified by design.
-/// - normal: Default visual state.
-/// - highlighted: Pressed/hovered/highlighted visual state.
-/// - disabled: Non-interactive. Same interaction as `loading`.
-/// - loading: Non-interactive with a centered spinner (uses `selectionColor` as tint).
-/// - confirmed: Success acknowledgment visual state.
-public enum RDSButtonState: Equatable {
-    case normal
-    case highlighted
-    case disabled
-    case loading
-    case confirmed
-}
-
-/// Immutable color palette used by `RDSButton`.
-/// `selectionColor` is applied to both text and icon.
-/// `underline` is used for text decoration (e.g., Primary.TextLink).
-public struct RDSButtonPalette: Equatable {
-    public let backgroundColor: Color
-    public let borderColor: Color
-    public let selectionColor: Color
-    public let underline: Bool
-    
-    /// Initializes a palette.
-    /// - Parameters:
-    ///   - backgroundColor: Color for the button background.
-    ///   - borderColor: Color for the 1pt border stroke.
-    ///   - selectionColor: Color for text and icon (and spinner tint in `loading`).
-    ///   - underline: Whether the title is underlined.
-    public init(
-        backgroundColor: Color,
-        borderColor: Color,
-        selectionColor: Color,
-        underline: Bool
-    ) {
-        self.backgroundColor = backgroundColor
-        self.borderColor = borderColor
-        self.selectionColor = selectionColor
-        self.underline = underline
-    }
-}
-
-/// Provides palettes for any (variant, type, state) combination.
-/// Supply your own to theme/brand the button without modifying the view.
-@MainActor
-public protocol RDSButtonPaletteProvider {
-    /// Returns the palette for a given configuration.
-    /// - Parameters:
-    ///   - variant: The visual variant.
-    ///   - type: The type within the variant.
-    ///   - state: The visual state.
-    /// - Returns: The resolved palette.
-    func palette(
-        for variant: RDSButtonVariant,
-        type: RDSButtonType,
-        state: RDSButtonState
-    ) -> RDSButtonPalette
-}
-
-// MARK: - Typography (Token-based)
-
-/// Provides the text style for each button size.
-/// By default it returns system fonts sized to match typical "buttonM" / "buttonS" styles.
-/// Replace the implementation to wire your `RDSTypography` tokens.
-public protocol RDSButtonTypographyProvider {
-    /// Returns the RDSUI `textStyleToken` to be used for a given button size.
-    /// - Parameter size: The RDSButtonSize.
-    /// - Returns: A `RDSTextStyleToken` to style the title text.
-    func textStyleToken(for size: RDSButtonSize) -> RDSTextStyleToken
-    
-    /// Whether the text should be scaled down slightly to prevent truncation on fixed widths.
-    /// - Parameter size: The RDSButtonSize.
-    func minimumScaleFactor(for size: RDSButtonSize) -> CGFloat
-}
-
 // MARK: - RDSButton View
 
 /// Design-system button with fixed sizes, explicit palettes per state, and optional icon placement.
@@ -213,56 +108,5 @@ public struct RDSButton: View {
                 .stroke(palette.borderColor, lineWidth: 1)
         )
         .clipShape(RoundedRectangle(cornerRadius: 2, style: .continuous))
-    }
-    
-    // MARK: - Content
-    
-    @ViewBuilder
-    private func content(palette: RDSButtonPalette) -> some View {
-        HStack(spacing: 0) {
-            if size == .iconLeft, let icon {
-                iconView(icon, color: palette.selectionColor)
-                Spacer().frame(width: 12)
-            }
-            
-            Text(title)
-                .rdsTextStyle(typographyProvider.textStyleToken(for: size))
-                .minimumScaleFactor(typographyProvider.minimumScaleFactor(for: size))
-                .foregroundStyle(palette.selectionColor)
-                .underline(palette.underline)
-                .lineLimit(1)
-            
-            if size == .iconRight, let icon {
-                Spacer().frame(width: 12)
-                iconView(icon, color: palette.selectionColor)
-            }
-        }
-        .padding(.vertical, 8)
-        .padding(.horizontal, 24)
-        .accessibilityElement(children: .combine)
-        .accessibilityAddTraits(.isButton)
-    }
-    
-    // MARK: - Helpers
-    
-    /// Renders the 20x20pt icon colored with `selectionColor`.
-    private func iconView(_ image: Image, color: Color) -> some View {
-        image
-            .resizable()
-            .scaledToFit()
-            .frame(width: 20, height: 20)
-            .foregroundStyle(color)
-            .accessibilityHidden(true)
-    }
-    
-    /// Fixed dimensions per size.
-    private static func dimensions(for size: RDSButtonSize) -> (width: CGFloat, height: CGFloat) {
-        switch size {
-        case .default:   return (135, 40)
-        case .large:     return (151, 56)
-        case .small:     return (108, 30)
-        case .iconLeft:  return (148, 40)
-        case .iconRight: return (148, 40)
-        }
     }
 }
